python/requirements.txt
#step001_blurandsave.py
python3
python3-pil
python3-numpy
